+++ server/auth/auth.go
@@ -184,6 +184,8 @@ func (r *Router) UserMiddleware(c *gin.Context, secret string, authURL string) {
 
 	c.Set(oauth.ContextAppKey, oauth.DriveAppID)
 
+	c.Set(oauth.ContextAppKey, oauth.DriveAppID)
+
 	c.Next()
 }